HOCS-5548: rework tests to use base class
Simplify the tests to utilise a base class - simplifying the test
pipeline for future extensions.
